cli: normalize plugin connect flow
Rust interface:
- `gestalt::commands::plugins::connect_with_browser_opener(&client, "acme_crm", Some("plugin"), None, |_| Ok(()))` still resolves a declared `_plugin` connection to the definition name on the wire.
- `gestalt::commands::plugins::connect_with_browser_opener(&client, "acme_crm", Some("plugin"), None, |_| Ok(()))` still preserves the raw `plugin` selector when the integration exposes no connection definitions.

stdin/stdout interface:
- `gestalt plugins connect manual-svc` -> select `Workspace OAuth` -> CLI prints `Connection: workspace` before continuing through the selected flow.
- `gestalt plugins connect widget_metrics` with multi-field manual auth now sends `{"credentials":{...}}` in the transport contract instead of collapsing distinct fields into a single prompt.

This splits the branchy connect logic into `cli/src/commands/plugins/connect.rs`, keeps disconnect alias normalization unchanged, and adds transport-level tests for selector aliasing, multi-field manual auth, and selected-connection auth override.
